moved patcher GUI into allolithe mid-way working on litheApp's OscillatorGUI
