+++ allolithe/allolithe/allolithe.hpp
@@ -3,6 +3,6 @@
 
 #include "allolithe/al_Node.hpp"
 #include "allolithe/al_SoundEngine.hpp"
-
+#include "allolithe/al_PatcherGUI.hpp"
 
 #endif //ALLOLITHE
