+++ litheApp/main.cpp
@@ -1,7 +1,6 @@
 // #include "LitheModules/LitheModules.h"
 #include "allolithe/allolithe.hpp"
 #include "Oscillator_GUI.hpp"
-#include "PatcherGUI.hpp"
 
 #define pr(thing) std::cout<<thing<<std::endl; /// Shorthand for printing a variable. 
 
@@ -10,10 +9,10 @@ int main()
 	Oscillator_GUI osc;
 	pr(Oscillator::moduleID);
 
-	PatcherGUI gui;
+	al::PatcherGUI gui;
 
-	// gui << osc;
-	// gui.openWindow();
+	gui << osc;
+	gui.openWindow();
 
     return 0;
 }
